+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - A host of binary vector operations:
   - `#dot_product` (aliased as `#inner_product` and `#scalar_product`)
   - `#angle`
-  - `#collinear?`, `#parallel?`, `#codirectional?`, `#opposite?`
+  - `#orthogonal?`, `#collinear?`, `#parallel?`, `#codirectional?`, `#opposite?`
   - `#scalar_projection`, `#vector_projection`
   - `#scalar_rejection`, `#vector_rejection`
 - Some utility methods related to linear subspaces: